// src/app/model/area/capitalhumano/Colaborador.ts
export class Colaborador{
    id_colaborador: number = parseInt(localStorage.getItem('id_colaborador')!);
    nombres: string = '';
    apellido_p: string = '';
    apellido_m: string = '';
    fecha_nacimiento: string = '';
    lugarNacimiento: string = '';
    sexo: string = '';
    id_nacionalidad: number = 0;
    tipoSangre: number = 0;
    donante: number = 0;
    alergias: string = '';
    enfermedades: string = '';
    limitaciones: string = '';
    camisa: string = '';
    txt_mailPersonal: string = '';
    nivel_estudio: number = 0;
    id_carrera: number = 0;
    certificado: number = 0;
    ingles: number = 0;
    CURP: string = '';
    RFC: string = '';
    NIMSS: string = '';
    foto: string = 'default.png';
    edo_civil: number = 0;
    celular: string = '';
    telefonoempresa: string = '';
    calleNumero: string = '';
    colonia: string = '';
    codigoPostal: string = '';
    dom_estado: string = '';
    municipio: string = '';
    pais: string = '';
    emergencia_nombre: string = '';
    emergencia_parent: string = '';
    emergencia_dom: string = '';
    emergencia_num: string = '';
    visa_num: string = '';
    pasaporte: string = '';
    poliza: string = '';
    registro_patronal: string = '';
    n_colaborador: string = '';
    nombreUsuario: string = '';
    email: string = '';
    tipo_empleado: number = 0;
    unidad: string = '';
    fecha_ingreso: string = '';
    jefe_inmediato: any;
    id_aprobador: any;
    colaborador_politicas: number = 0;
    id_region: number = 0;
    id_sucursal: number = 0;
    id_area: number = 0;
    id_departamentos: number = 0;
    id_geo: number = 0;
    id_puesto: any;
    especialidad: number = 0;
    desempeno: number = 0;
    tarjeta: string = '';
    clabe: string = '';
    tabulador_id: number = 0;
    tab_nivel: number = 0;
    modif_salarial: string = '';
    porcentaje: string = '';
    pass: string = '';
    passNuevo: string = '';
    passNuevo2: string = '';
    infonavit: number = 0;
    fonacot: boolean = false;
    nr_reporte: string = '';
    monto_sgmm2: number = 0;
    monto_sgmmenor2: number = 0;
    poliza_sgmmenor2: string = '';
    banco: string = '';
    geografica_sap: string = '';
    ingreso_bruto_men: number = 0;
}